Add toastr messages when a course is saved or API errors when trying to save.
Put an async/await alternative to a function which user .then(...).catch(...). It worked but commented it out as eslint complains and get a black error in browser every time save file. Refreshing the page resolves things but will have to wait until using a different project configuration if don't want to have to put up with this.