Add webpack-4-test folder containing 2 files edited when tried upgrading from webpack v1 to v4.
A number of errors occurred when ran "npm start". Although the webpack.config.dev.js was adjusted so we could actually build the code, the page did not load. It is unlikely to be a good idea to try to upgrade all the packages from within this project. We should probably find a more modern template.