Add PageNotFound component which is displayed if the course ID in the URL cannot be found.
The current solution looks hackish. It pushes the page-not-found-url to browserHistory if course does not exist in mapStateToProps() of ManageCoursePage. However, pushing to browserHistory des not appear to stop the component from continuing on its lifecycle and, although we see the correct outcome in the browser, we also see at least one error within the console.